Apply un-substitution in reverse dependency order.
This doesn't work satisfactory yet. For example, in the expansion of V,
there shows up a T, despite V does not depend on T. And so V is not
recognized anymore. 
At the end of the day 'unsubst' depends on the shape of the tree for
combinators that do not depend on each other.